logic for total with quantity and use float for quantity

+++ src/main/java/com/order/models/Food.java
@@ -3,7 +3,7 @@ package com.order.models;
 public class Food {
 
     private String food;
-    private int quantity;
+    private Float quantity;
     private Float cost;
 
     public String getFood() {
@@ -14,11 +14,11 @@ public class Food {
         this.food = food;
     }
 
-    public int getQuantity() {
+    public Float getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(Float quantity) {
         this.quantity = quantity;
     }
